refactor(matrix): Extract perf test utils into their own module
